refactor(test): extract the LISH protocol stream harness
